# Improve Base64 decoding performance