--- a/src/base64.cr
+++ b/src/base64.cr
@@ -312,7 +312,7 @@ module Base64
   private macro decode_chunk(raise_error, chunk_pos, bytes, &block)
     %buffer = 0_u32
     {% for byte, i in bytes %}
-      %decoded = DECODE_TABLE[ {{byte}} ]
+      %decoded = DECODE_TABLE.unsafe_fetch({{byte}})
       %buffer = (%buffer << 6) + %decoded
       {% if raise_error == true %}
         raise Base64::Error.new("Unexpected byte 0x#{{{byte}}.to_s(16)} at #{{{chunk_pos}} + {{i}}}") if %decoded == 255_u8