--- a/src/base64.cr
+++ b/src/base64.cr
@@ -266,13 +266,11 @@ module Base64
       end
       break if bytes > fin
 
-      decode_chunk(
+      yield_decoded_chunk_bytes(
         raise_error: true,
         chunk_pos: bytes - bytes_begin,
         bytes: {bytes[0], bytes[1], bytes[2], bytes[3]}
-      ) do |byte|
-        yield byte
-      end
+      )
       bytes += 4
     end
 
@@ -287,21 +285,17 @@ module Base64
     when 1
       raise Base64::Error.new("Wrong size")
     when 2
-      decode_chunk(
+      yield_decoded_chunk_bytes(
         raise_error: true,
         chunk_pos: bytes - bytes_begin,
         bytes: {bytes[0], bytes[1]}
-      ) do |byte|
-        yield byte
-      end
+      )
     when 3
-      decode_chunk(
+      yield_decoded_chunk_bytes(
         raise_error: true,
         chunk_pos: bytes - bytes_begin,
         bytes: {bytes[0], bytes[1], bytes[2]}
-      ) do |byte|
-        yield byte
-      end
+      )
     end
   end
 
@@ -309,7 +303,7 @@ module Base64
   # The first argument (raise_error) determines if the macro raises an error when an invalid character is processed.
   # The second argument (chunk_pos) is only used for the resulting error message.
   # The resulting bytes are then each yielded to the given block, using the variable `byte`.
-  private macro decode_chunk(raise_error, chunk_pos, bytes, &block)
+  private macro yield_decoded_chunk_bytes(raise_error, chunk_pos, bytes, &block)
     %buffer = 0_u32
     {% for byte, i in bytes %}
       %decoded = DECODE_TABLE.unsafe_fetch({{byte}})
@@ -321,8 +315,7 @@ module Base64
 
     # Each byte in the buffer is shifted to rightmost position of the buffer, then casted to a UInt8
     {% for i in 2..(bytes.size) %}
-      byte = (%buffer >> {{ (4 - bytes.size) * 2 + (8 * (bytes.size - i)) }}).to_u8!
-      {{ yield }}
+      yield (%buffer >> {{ (4 - bytes.size) * 2 + (8 * (bytes.size - i)) }}).to_u8!
     {% end %}
   end