--- a/src/base64.cr
+++ b/src/base64.cr
@@ -288,9 +288,8 @@ module Base64
   end
 
   # This macro decodes the given chunk of (2-4) base64 characters.
-  # The first argument (raise_error) determines if the macro raises an error when an invalid character is processed.
-  # The second argument (chunk_pos) is only used for the resulting error message.
-  # The resulting bytes are then each yielded to the given block, using the variable `byte`.
+  # The argument chunk_pos is only used for the resulting error message.
+  # The resulting bytes are then each yielded.
   private macro yield_decoded_chunk_bytes(*bytes, chunk_pos)
     %buffer = 0_u32
     {% for byte, i in bytes %}