Supersedes #11032 

This PR improves the performance of Base64 decoding.
It replaces the previous PR to ease the reviewing of this PR.
It does not include the bang variants.

Benchmarks:
```
Benchmark: base64 decode (8 characters)
old  48.03M ( 20.82ns) (± 1.74%)  16.0B/op   1.06× slower
new  50.82M ( 19.68ns) (± 1.78%)  16.0B/op        fastest

Benchmark: base64 decode (8000 characters)
old 180.71k (  5.53µs) (± 1.34%)  5.88kB/op   1.17× slower
new 211.70k (  4.72µs) (± 1.24%)  5.88kB/op        fastest

Benchmark: base64 decode (80000000 characters)
old  20.35  ( 49.13ms) (± 0.30%)  57.2MB/op   1.19× slower
new  24.31  ( 41.13ms) (± 1.08%)  57.2MB/op        fastest
```